operator/requirements.txt
python-jose[cryptography]
httpx
kopf
kubernetes
pytz